chore: move Arinc429Word out of React context (#6127)
* Move Arinc429Word out of React context

* Fix linting errors

* Add convenience functions

* Fix linting error
