+++ src/instruments/src/PFD/AttitudeIndicatorFixed.tsx
@@ -1,4 +1,4 @@
-import { Arinc429Word } from '@instruments/common/arinc429.js';
+import { Arinc429Word } from '@shared/arinc429';
 import React from 'react';
 import { getSimVar } from '../util.js';
