+++ src/instruments/src/PFD/DigitalAltitudeReadout.tsx
@@ -1,4 +1,4 @@
-import { Arinc429Word } from '@instruments/common/arinc429';
+import { Arinc429Word } from '@shared/arinc429';
 import React from 'react';
 
 const TensDigits = (value: number, offset: number, color: string) => {
