+++ src/instruments/src/PFD/SpeedIndicator.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { Arinc429Word } from '@instruments/common/arinc429';
+import { Arinc429Word } from '@shared/arinc429';
 import { VerticalTape, BarberpoleIndicator } from './PFDUtils';
 import { getSimVar } from '../util.js';
