+++ src/instruments/src/PFD/VerticalSpeedIndicator.tsx
@@ -1,4 +1,4 @@
-import { Arinc429Word } from '@instruments/common/arinc429';
+import { Arinc429Word } from '@shared/arinc429';
 import React from 'react';
 
 interface VerticalSpeedIndicatorProps {
